+++ atoum/admin/product.py
@@ -1,4 +1,7 @@
 from django.contrib import admin
+from django.urls import path, reverse
+from django.http import HttpResponseRedirect, QueryDict
+from django.utils.translation import gettext_lazy as _
 
 from import_export import resources
 from import_export.admin import ImportExportModelAdmin
@@ -8,6 +11,7 @@ from smart_media.admin import SmartAdminMixin
 
 from ..forms import ProductAdminForm
 from ..models import Category, Product
+from ..views.admin import ProductActionMoveAdminView
 
 
 class ProductResource(resources.ModelResource):
@@ -54,6 +58,45 @@ class ProductAdmin(SmartAdminMixin, ImportExportModelAdmin):
         "brand",
     )
     resource_classes = [ProductResource]
+    actions = ["move_to_category"]
+
+    def get_urls(self):
+        """
+        Set some additional custom admin views
+        """
+        urls = super().get_urls()
+
+        extra_urls = [
+            path(
+                "actions/move/",
+                self.admin_site.admin_view(
+                    ProductActionMoveAdminView.as_view(),
+                ),
+                name="atoum_admin_product_move",
+            ),
+        ]
+
+        return extra_urls + urls
+
+    @admin.action(description=_("Move selected products to a category"))
+    def move_to_category(self, request, queryset):
+        """
+        Redirect to dedicated form view.
+
+        Object id of selected products along admin changelist filters are passed into
+        redirection.
+        """
+        selected = queryset.values_list("pk", flat=True)
+        url = reverse("admin:atoum_admin_product_move")
+
+        # Ensure changelist filters from GET request are passed to the view
+        request_query = QueryDict(mutable=True)
+        request_query.update(request.GET.dict())
+        request_query.update({"selection": ",".join(str(pk) for pk in selected)})
+
+        return HttpResponseRedirect(
+            url + "?{}".format(request_query.urlencode(safe=","))
+        )
 
 
 class ProductAdminInline(admin.StackedInline):