+++ atoum/form_helpers/__init__.py
@@ -1,8 +1,9 @@
 from .default import DefaultFormHelper
 from .search import AdvancedSearchFormHelper
-
+from .admin import AdminProductFormHelper
 
 __all__ = [
     "AdvancedSearchFormHelper",
     "DefaultFormHelper",
+    "AdminProductFormHelper",
 ]