+++ atoum/forms/__init__.py
@@ -1,5 +1,5 @@
 from .category import CategoryAdminForm
-from .product import ProductAdminForm
+from .product import ProductAdminForm, ProductMoveToAdminForm
 from .search import GlobalSearchForm
 from .shopping import ShoppingAdminForm, ShoppingItemInlineForm
 
@@ -8,6 +8,7 @@ __all__ = [
     "CategoryAdminForm",
     "GlobalSearchForm",
     "ProductAdminForm",
+    "ProductMoveToAdminForm",
     "ShoppingAdminForm",
     "ShoppingItemInlineForm",
 ]