+++ atoum/views/admin/__init__.py
@@ -1,0 +1,6 @@
+from .product import ProductActionMoveAdminView
+
+
+__all__ = [
+    "ProductActionMoveAdminView",
+]